+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to this project will be documented in this file.
 
+## 📍 [1.5.4] - 2025-07-20
+
+- added: Unit second on "Base Time" (Skill Node)
+- added: Unit second on "Casting Time" (Skill Node)
+- added: Unit second on "Cooldown" (Skill Node)
+- updated: Resize Tooltip to suit each screen size (Skill Node)
+- updated: Skill Description (add parameter "STR Scaling")
+- updated: Skill Description (add parameter "STA Scaling")
+- updated: Skill Description (add parameter "INT Scaling")
+- updated: Skill Description (add parameter "DEX Scaling")
+- changed: text-md to text-base (Tailwind CSS)
+- changed: Decrease icon size 16px to 14px (Skill Node)
+- fixed: UI icon size margin on Mobile device (Skill Node)
+- fixed: Spacing of description "Spell Range" (Skill Node)
+- fixed: Skill Parameter description "Sword Attack" (Skill Node)
+- fixed: Skill Parameter description "Axe Attack" (Skill Node)
+- fixed: Skill Parameter description "Hit Rate" (Skill Node)
+- fixed: Skill Parameter description "Element Attack" (Skill Node)
+- fixed: ALl langauge of stat "STR/STA/INT/DEX" (Skill Node)
+
 ## 📍 [1.5.3] - 2025-07-18
 
 - added: Update Version Dialog